Add GNU readline support for user input
This commit adds import for the readline module at the top of
terminal_interface.py to provide GNU readline support for user
input. This will allow Unix users to edit their input and browse their
command history like in a UNIX shell. The readline module is imported
conditionally, so the script does not break for Windows users where
this module is typically not available.

pyproject.toml already mentions the ability to do this when declaring
the pyreadline dependency and interpreter.py also attempts importing
readline, but it doesn't work unless the import is in the same module
using the input function.